Just updated the game adding a title screen.
Functionality of the buttons no have been updated along with health. There are more debugging buttons for the mean time.
FUTURE UPDATE- The main player class will be updated to allow more than one main character. Have to add the width
and heigh of the sprite and then layout the sprite sheet according to number of animations.
This'll be generalized just like the enemy class file.